requirements.txt
asgiref==3.2.10
Django==3.1.1
django-allauth==0.44.0
django-countries==7.1
django-filter==2.4.0
django-js-asset==1.2.2
django-model-utils==4.1.1
docutils==0.17
gunicorn==20.1.0
oauthlib==3.1.0
parsel==1.6.0
Pillow==8.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.3.1